hjh update

+++ src/main/java/com/lotteon/dto/product/ProductDTO.java
@@ -41,7 +41,7 @@ public class ProductDTO{
     private Boolean isSaled; // 판매가능여부
     private String sellerId;
     private String productCode;
-    private int hit;
+    private long hit;
 
     private String file190;
     private String file230;
